fix(ios): use Float not double for selection rect emit
Codegen types OnChangeSelection's rect fields as Float (from the spec's
Float annotation). The designated-initializer emit cast to double, which
is a narrowing conversion in C++ brace-init and fails to compile. Cast
to Float to match the struct.

+++ ios/EnrichedTextInputView.mm
@@ -2181,10 +2181,10 @@ static UIColor *katavParseHexColor(NSString *hex) {
          .end = static_cast<int>(textView.selectedRange.location +
                                  textView.selectedRange.length),
          .text = [textAtSelection toCppString],
-         .rectX = static_cast<double>(selectionRect.origin.x),
-         .rectY = static_cast<double>(selectionRect.origin.y),
-         .rectWidth = static_cast<double>(selectionRect.size.width),
-         .rectHeight = static_cast<double>(selectionRect.size.height)});
+         .rectX = static_cast<Float>(selectionRect.origin.x),
+         .rectY = static_cast<Float>(selectionRect.origin.y),
+         .rectWidth = static_cast<Float>(selectionRect.size.width),
+         .rectHeight = static_cast<Float>(selectionRect.size.height)});
   }
 
   // manage selection changes
